tests/requirements.txt
selenium==4.16.0
pytest==7.4.3
pytest-html==4.1.1
pytest-timeout==2.2.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
faker==21.0.0
psycopg2-binary==2.9.9
